Remove most globals from tkp/database/database.py
database.py contained a bunch of unnecessary globals,
defining e.g. default database name.
Since these are defined at import time, later alterations
to the config have no effect on the defaults.
Now, the defaults are always in sync with the config dict.

The remaining global, ENGINE, is valid, as it records which
module has been imported as the database engine.