Fix for Issue #3367 - correct db insertion of image.tau_time.
NB This commit will only work in conjunction with the database commit 8ed6929

I've updated the relevant image insertion routines so we can now insert
all relevant parameters into the database.

Note this commit also includes some modifications to the unit test decorators,
I've tweaked it so the "requires_database" test decorator now switches the
database over to the 'test_database' defined in config.

(This was my original plan but I had to get rid of the globals in the
 database init routines first!)