Strip down subpackage tkp.database to the bare essentials.
There were a lot of modules in tkp.database that clearly contain useful code,
but are not integral to the workings of the Trap in its current state.

I've moved them to a folder of their own for the time being.
I imagine eventually we will want to reintegrate some of the code into
the tkp package, perhaps under a 'plots' or 'quality' subpackage.

I've also deleted gsm.py gsmutils.py as they duplicate the (newer)
versions in the LOFAR/CEP repository.