Include database dir in CMakeLists, pull across database CMakeLists.
These got left behind in the repo merge.
Possibly it is often sensible to not use them and just install from the repo,
but my build scripts expect an installed copy of the setup script,
so they can put some in handy symlinks to the catfiles.

Also renamed setup -> setup.sh, changed default database to 'testdb'.