Update MonetDB log suppression, output default datapath if used.
If a database error is encountered, the MonetDB log spews out hundreds
of lines of output.
While this may be useful for tricky bugs, it obscures the source of
the bug if you've made a simple syntax error, as you must scroll
up many pages to see the test that failed.

By changing the logger level in the runtests wrapper, we suppress the output
so we can just see the unit tests error messages.