Limit area over which debug images are made
When calculating source and residual images, we previously calculated an
elliptical gaussian for each source over all pixels in the output image. Given
a large image size, this is very slow. Further, beyond a few FWHMs from the
centre of the source, the flux is very close to 0.

Here, we limit the fraction of the image over which each gaussian is
evaluated. This t pisloses effectively no flux (I think in the worst case of a
perfectly circular source we lose 1.9e-7%!) and is significantly faster.

See issue #3770.