Merging tests to avoid repeated setUps()
The setUp() method here performs source extraction on a large image, which is
computationally expensive. And it was doing the same source extraction, with
the same parameters, on the same image, for each of 6 different tests. By
merging those tests into one, we only need to run setUp() once, and things
should get very much faster.