Revert "remove depricated scripts"
This reverts commit da03737c8e7aacaa2f6c8e6a28e3c6a9eee0d89f.

This broke the build, since the deprecated scripts are still referred to
in eg setup.py. Reverting rather than just fixing up setup.py since
whether they should actually be included or not still seems to be
unclear, and I am not taking a position on that (see discussion at
https://github.com/transientskp/trap/pull/29).

Should fix #3894.