changed the way queries are split
OK, this is dirty hack for now. The problem is that
the python API of MonetDB can't parse multiple resultsets. This
happens when you execute multiple queries in one run. The upgrade
script splitted statements by ';', which doesn't work with
functions. So for now we have to split functions by adding %SPLIT%
after every query.