Iss4084, add meridian wrapping
When the extraction region of an image crosses the ra 360/0 meridian,
the source association runs a slightly modified query.

For the meridian-wrap query we have to calculate the average values slightly differently.
Also added circular search area around source for the assoc queries (was a box)

A unittest is being added to test_association.

Note that have not yet processed any merdian-cross images.