Give helpful error messages if celery-pipeline cannot find config files.
This sort of stuff was taken care of by cuisine, we need to do it ourselves
here. Otherwise a missing file (or execution from wrong directory) does not
raise an exception, it just returns an empty config object. This results in
confusing errors about 'section not found in config file' or similar.