My very own parser-shaped wheel (with apologies to John).
An attempt to stick to the principle that each parameter should be defined
"Once and Only Once," through use of ast.literal_eval. No 'getInt's here,
thankyou!

This way, the type of the parameter is defined by how it is written in
the parset, in the same way as standard python syntax, e.g. ``1`` is an int,
but ``1.`` is a float. Strings should be enclosed in quotes, but we
failsafe to converting anything we don't recognise to a string anyway.

I've also reworked all the step tests to actually employ the default
parameterset files as their test parameters. This avoids repetition
and ensures the defaults are sensible / in sync with the functions
they supply.

Working with celery and local.

p.s. Wheee, python!