Accessors refactoring.
Make use of header info to determine data type, and hence expected metadata.
Raise an IOerror exception at load-time if expected metadata is not found.
Add Lofar data accessors subclasses for both Fits and Casatable.

NOTE: For both Fits and Casa formats, if the beam is specified it should
be specified *in degrees*. This makes the most sense, since pixel-size
in the current image is not known ahead of time.