Fix for #4497.
In the case of a sky region which crosses the meridian, we were adding 180
degrees to avg_wra to make the calculations tractable, but then forgetting to
subtract it again afterwards.

Normally, this isn't a problem, since avg_wra isn't used in the calculations
for regions which straddle the meridian. But if that source was then seen in a
region which doesn't straddle the meridian, the incorrect value of avg_wra
causes its mean RA (wm_ra) to be 180 degrees wrong.