Change back to initial working dir at end of test
This stops a segfault occuring in future tests which use pyrap, which
apparently really doesn't like starting with a non-existent working dir (and
exits with a ``glibc detected *** /var/lib/jenkins/virtualenvs/tkp/bin/python:
munmap_chunk(): invalid pointer``).

Probably we should check all uses of temp dirs in the tests for the same
issue, but I think this fixes the error that's currently blocking Jenkins.