Subtract one from CRPIX1/2 in FITS headers
This applies when setting up our WCS object.

FITS images are defined such that the data array starts at (1,1). Most other
arrays (including NumPy, CASA tables, etc) are conventionally indexed from
(0,0). The Trap code makes the assumption that all images start at (0,0).

This is just a matter of definition, *except* insofar as it affects the WCS
reference pixel (CRPIX). That must be shifted by one pixel to account for the
difference in definition in Trap images.

You can see this happening in reverse by taking a CASA image and checking
coords/direction0/crpix, then running it through image2fits and checkint the
CRPIX1 and CRPIX2 headers -- the latter should be incremented by one with
respect to the former.

This resolves #4770.