Push error_radius to database
Source association is based on on-sky distance uncertainties.
The error radius as provided by source finder is propagated
to the database, where the source's positional uncertainty
is based on the error radius and the systematic errors.

Needed in the De Ruiter radius calculations
Adjusted unittests for error_radius
Increment db version number