More robust NCP tests.
When a source is measured at ra=0.0 ('exactly') for a sequence of images, the
weighted ra averages are numerically fluctuating around 0.0. Due to the nature
of floating point arithmetic, an RA in the range [-8e-14, 0) + 360 == 360,
rather than 359.999. We do not regard an RA of 360 as meaningful, and
therefore force it to 0.0.