Changes to accomodate errors on ra
We introduce new db columns:
uncertainty_ew: the source positional on-sky uncertainty in the
east-west direction
uncertainty_ns: the source positional on-sky uncertainty in the
north-south direction,
ra_err and decl_err are the 1-sigma errors on ra and decl, resp.

Now we have an average of ra_err, which is the error on ra
(decl-dependent), whereas wm_uncertainty_ew is the weighted mean of
the on-sky error in the east-west direction, along ra.

Note that the db stores ra_err, uncertainty_ew, ra_fit_err in degrees
and error_radius and ra_sys_err in arcsec (same for decl errors).

Using on-sky errors (uncertainty_ew & _ns) in weighted averages
in runningcatalog, and in De Ruiter calculations.

Increment db version number

It is known, however, that the source associations are not always
handled correctly for sources at high declination and large RA values.
Issue 4934 will solve that.