Eliminate cast to Numeric and use FLOOR()
Postgres doesn't provide round(double, int): see
<https://stackoverflow.com/questions/13113096/how-to-round-an-average-to-2-decimal-places-in-postgresql>
for details.

Casting to NUMERIC limits the dynamic range, particularly with Monet which
only supports a maximum of 18 sig figs. There's no point in doing this when
all we want to do is round. Here we use floor() instead.

Note that the change to the stored function mandates a schema version bump.