Overhaul configuration - parse job_params to a single dict.
Also move some subroutines out of the main trap logic module.

In hindsight, the way I had the parsing code set up was not good:
- parsing per config section is just a messy reflection of how we
  work with ConfigParser, and results in many separate config dicts
  floating around.
- We never used the 'dict to ConfigParser' functionality, and now
  that I think about it that mapping is not well defined -
  ConfigParser objects have defaults that are
  not preserved when we extract the resulting section values.

Now, one effect is that when we dump a record of the pipeline
configuration to file, we dump it in dict format rather than
ConfigParser format. This means you cannot easily feed it
into a new run, but I think that's a negligible and necessary loss -
we don't want to record that some dir in the pipeline cfg points to %(cwd)s,
we want to record the *actual directory path that resolved to*.