Rework unit tests for checking flux catalogs and per-timestep var-indices.
I've tried to eliminate any 'magic number' comparison, everything is
cross-calculated from the sources inserted. This means you can
freely change those source parameters, and so long as the relative
positions are maintained such that associations remain the same,
the tests should just work.

Also, I altered the way in which we retrieve per-timestep data,
to work on a per-runcat-entry basis. Grabbing *everything* related
to the current dataset seems difficult to verify.

On a sidenote, test_lightcurve is pretty crufty and I was tempted to
delete it rather than refactor it. However I guess it has marginal
value in testing the fairly simple case of several one-to-one
associations being carried through multiple images.