Improved MockSource class for unit-testing transient logic.
We now create a lightcurve based on datetimes (lots of future potential here
for testing against realistic lightcurve sims), and then test the resulting
source against the skyregion and extraction RMS / threshold to determine
if we see this source in this image.

I've also added some updated, basic unit-tests for the transients-detection
logic - these test two simple cases of a new source and a variable source.