Clean out unused code.
This removes all the material relating to storing and processing external
catalogues as part of the pipeline database and to feature extraction based on
lightcurves.

None of this was in use. Some of it may be resurrected in the future, but the
relevant subsystems need design and spec work before coding.