DB associations: Some more renaming and some re-ordering.
I find this logic layout easier to reason about and follow.
I've updated the docs section accordingly, there's now much
less duplicated information - the details should always be
in the docstring, and preferably only in the docstring.

Finally, I've made the 'delete-1-to-many' code more uniform,
which may or may not be marginally more robust / faster.