Implement deconvolution in Cython.
This replaces the "deconv.so" module based on Fortan code lifted from AIPS
with a home-grown version using Cython.

This enables us to remove two pieces of code which we don't hold copyright on
and which are licensed under (viral) GPL licenses, thereby paving the way for
issue #4712.

For reference, this Cython code achieves almost exactly the same performance
as the Fortran code. Switching it to use floats (rather than doubles) actually
makes it significantly faster, at the cost of somewhat reduced accuracy. Store
that one away for a rainy day when performance is a big issue...