Use pywcs to handle world coordinate systems.
This replaces the old python-wcslib wrapper (inherited from the LOFAR User
Software Group in days of yore) with pywcs.

This enables us to remove a piece of code on which we don't hold copyright and
which is completely lacking in licensing information, thereby paving the way
for issue #4712.

Note this removes the (remnants of) the ability to dynamically change the
equinox of Detection objects, but that functionality was unused and partially
broken.