Monitoring sources uniquely associated
Monitoring sources are now entered into the runningcatalog table
at the start of a pipeline run. There is no need for an extra table.
Monitoring sources are recognised as such by the boolean column
``mon_src``. Those that do not fall inside any of the images and thus
cannot be forced fit, do not have a ``xtrsrc`` id. In the case they do,
the ``xtrscr`` id is set to its first occurence, just like for the
normal detections. As a consequence this column is now defined as
nullable.
Only monitoring sources are added that do not yet exist in the runcat
table (in case of rerunning with dataset id = -1).

During pipeline processing, it checks for existing monitoring sources
that fall within the current image, by checking image skyregion and the
monitoring source co-ordinates.

The monitoring-source associations follow a simplified version of the
normal association procedure. The runcat and runcat_flux tables are updated
with the forced-fit values, and ``xtrsrc`` is when it is fit for the first
time.

All associations are 1-to-1, and of type 8 (including the first).

Due to the changes for this issue, the RhombusError now appears at the
``_insert_1_to_many_runcat_flux()`` function in the normal association
procedure.

Unittests and documentation updated.