Update SQL queries to be compatible with MonetDB Oct2014-SP4
'First CASE, then SQRT':
MonetDB throws an error if we attempt to take the square root of
a number very close to zero. We actually already assign sensible
values in those cases via the CASE statement, but we need to
ensure that the CASE evaluation completes before SQRT is
evaluated.