Replace 'abstractproperty' usage in DataAccessor with dedicated metaclass.
Created a metaclass, 'RequiredAttributes', which allows you to define a list
of required attributes and check that they are defined and not-None, post
initialization.

To see how this works in practice, add an extra (dummy) attribute
to the _required_attributes list of either the DataAccessor or
LofarAccessor classes, then try running the accessor unit tests.
You should get an error message complaining that it's not defined,
and reporting the full list of attributes required to instantiate
each accessor class tested.

This results in much cleaner code than use of @abstractproperty, but note
that in it's current incarnation we lose one possibly useful feature -
the abstract properties were read-only, while these are just regular
attributes. I don't think this is a problem, but if anyone feels
strongly about it perhaps I can tinker with the metaclass some more until
that's replaced too (replace all the _requiredattributes with property
getters post initialization?).

I also took a bit of Gij's advice and made the data/header parsing functions
inheritable, so they can be re-used down the inheritance stack.
However, instead of making them regular class functions, I've made them static
functions - this means that they have to return their values and makes it
absolutely clear in the __init__ function which attributes are being
initialized by which function (my worry with the alternative is that you end
 up having to search through the whole inheritance stack to find the relevant
function which has the side-effect of setting a particular attribute).