FIx inject.py file check
inject.py was never finding the file (a CASA dir) which was present when run, which I'm pretty sure was because of line 99.

In this case, if the target_file is a dir then it will return True as it's missing a not. If the 'or' is changed to 'and not' then I believe the logic is correct and will only be flagged when the target_file is not found completely. There might be a neater way to put it?