Sourcefinder: Crash out early if image-data is uniform (flat).
NB. removed an line of code which masked any pixels with the value 0.
This was causing issues because it was masking all pixels in the synthetic
test-image, and so the min == max check was coming up false.
It makes sense to remove that line anyway, since there's no reason
why a pixel should be zero-valued in the real world and still be valid
data.